move DirichletProcess to edward/models/
