+++ docs/tex/api/model.tex
@@ -83,6 +83,8 @@ detailed below.
 .. autoclass:: edward.models.RandomVariable
    :members:
 
+.. autoclass:: edward.models.DirichletProcess
+
 .. autoclass:: edward.models.Empirical
 
 .. autoclass:: edward.models.PointMass
