+++ docs/tex/api/reference.tex
@@ -12,6 +12,7 @@ available:
 {%sphinx
 
 * :class:`edward.models.RandomVariable`
+* :class:`edward.models.DirichletProcess`
 * :class:`edward.models.Empirical`
 * :class:`edward.models.PointMass`
 * {{tensorflow_distributions}}
