+++ edward/models/__init__.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+from edward.models.dirichlet_process import *
 from edward.models.empirical import *
 from edward.models.point_mass import *
 from edward.models.random_variable import *
